img search shhutterstock added

+++ api/service/imgService.js
@@ -50,6 +50,30 @@ function deleteFile(filename){
     });
 };
 
+function increaseImgsCounter(imgs){
+    var index = 0;
+    return new Promise((resolve, reject)=>{
+        if(imgs.length === 0)
+            return resolve();
+        imgs.forEach((img, index)=>{
+                    Img.findOne({ 'hash': img.hash}).exec().then(doc=>{
+                        if(doc){
+                            doc.timesUsed++;
+                            doc.save((err)=>{
+                                if(err)return reject(err);
+                                index++;
+                                if(index === imgs.length)
+                                    return resolve();
+                            });
+                        }
+                        else{
+                            return reject("image does not exist");
+                        }
+                    })
+                })
+        })     
+}
+
 
 function saveImgDb(filename, hash, contentType){
     return new Promise((resolve, reject)=>{
@@ -57,7 +81,7 @@ function saveImgDb(filename, hash, contentType){
             if(img){
                 img.timesUsed++;
                 img.save((err)=>{
-                    if(err)reject(err);
+                    if(err)return reject(err);
                     return resolve(hash);
                 });
                 return;
@@ -82,15 +106,15 @@ function saveImgDb(filename, hash, contentType){
 }
 
 
-function downloadArray(urls, userId, callback){
+function downloadArray(imgs, userId, callback){
     return new Promise((resolve, reject)=>{
         var imgHashes = [];
         var contentType;
-        if(!urls || urls.length === 0)
+        if(!imgs || imgs.length === 0)
             return resolve([]);
-        urls.forEach((url)=>{
-            const imgPath = imgDir + "/" + getImgName(url, userId);
-            download(url, imgPath)
+        imgs.forEach((img)=>{
+            const imgPath = imgDir + "/" + getImgName(img.url, userId);
+            download(img.url, imgPath)
             .then(ct=>{
                 contentType = ct;
                 return md5File(imgPath);
@@ -99,16 +123,20 @@ function downloadArray(urls, userId, callback){
                  return saveImgDb(imgPath, hash, contentType);
                     })
             .then(hash=>{
-                     imgHashes.push(hash);
+                     imgHashes.push({
+                         hash: hash,
+                         width: img.width,
+                         height: img.height
+                     });
                      return deleteFile(imgPath);
                     })
             .then(()=>{
-                    if(imgHashes.length === urls.length)// if satisfies condition, this is the last cycle of the loop
+                    if(imgHashes.length === imgs.length)// if satisfies condition, this is the last cycle of the loop
                         return resolve(imgHashes);  
                     })
                     .catch(err=>{
                         logger.error(err);
-                        return reject({success:false, msg:String(err)});
+                        return reject(String(err));
                     });
         });
     });
@@ -121,7 +149,17 @@ function deleteImgOnce(hash, callback){
                      .then(img=>{
                         img.timesUsed--;
                         if(img.timesUsed <= 0){
-                            AWSService.removeFromS3(hash, callback);
+                            img.remove((err, count)=>{
+                                if(err){
+                                    logger.error(err);
+                                    return callback({success:false, msg:String(err)});
+                                }
+                                if(count === 0){
+                                        logger.error("could not delete image(hash:"+hash +") from card");
+                                        return callback({success:false, msg:"could not delete image from card"}); 
+                                }
+                                AWSService.removeFromS3(hash, callback);
+                            });
                         }
                         else{
                             img.save((err)=>{
@@ -139,15 +177,15 @@ function deleteImgOnce(hash, callback){
 
 
 
-function deleteImgsOnce(imgsId){
+function deleteImgsOnce(imgs){
     return new Promise((resolve, reject)=>{
-        if(!imgsId || imgsId.length === 0)
+        if(!imgs || imgs.length === 0)
             return resolve(true);
-        imgsId.forEach((id, index)=>{
-            deleteImgOnce(id, (err)=>{
+        imgs.forEach((img, index)=>{
+            deleteImgOnce(img.hash, (err)=>{
                 if(err)
                     return reject(err);
-                if(index === imgsId.length - 1)
+                if(index === imgs.length - 1)
                     return resolve(true);
             });
         });
@@ -157,5 +195,6 @@ function deleteImgsOnce(imgsId){
 module.exports = {
     downloadArray: downloadArray,
     getImg: AWSService.getImgFromS3,
-    deleteImgsOnce: deleteImgsOnce
+    deleteImgsOnce: deleteImgsOnce,
+    increaseImgsCounter: increaseImgsCounter
 }
